requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
decorator==4.4.2
easydict==1.9
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
flask-serialize==1.1.9
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
idna==2.9
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
newsapi-python==0.2.6
psycopg2-binary==2.8.5
pycparser==2.20
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.5
urllib3==1.25.9
validators==0.15.0
Werkzeug==1.0.1
WTForms==2.3.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4
